# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive API for managing cargo and vessel tracking in logistics operations",
        "title": "DEUS Logistics API",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/cargoes/{id}/tracking": {
            "get": {
                "description": "Retrieve the complete tracking history for a cargo shipment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracking"
                ],
                "summary": "Get cargo tracking history",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Cargo ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tracking history retrieved",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/http.TrackingEntryResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid cargo ID format",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Cargo not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new tracking entry for a cargo shipment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracking"
                ],
                "summary": "Add a tracking entry",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Cargo ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tracking entry payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.AddTrackingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tracking entry created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.TrackingEntryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Cargo not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "http.AddTrackingRequest": {
            "type": "object",
            "required": [
                "location",
                "status"
            ],
            "properties": {
                "location": {
                    "type": "string",
                    "example": "Port of Mumbai"
                },
                "note": {
                    "type": "string",
                    "example": "Passed customs inspection"
                },
                "status": {
                    "type": "string",
                    "example": "in_transit"
                }
            }
        },
...
        "http.TrackingEntryResponse": {
            "type": "object",
            "properties": {
                "cargo_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174002"
                },
                "location": {
                    "type": "string",
                    "example": "Port of Mumbai"
                },
                "note": {
                    "type": "string",
                    "example": "Passed customs inspection"
                },
                "status": {
                    "type": "string",
                    "example": "in_transit"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-15T12:30:00Z"
                }
            }
        },
...
        "response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                }
            }
        }
    }
}